Address CR comments
Changes:
  - Moved a few test related code to `test_utils.rs` file
  - Updated benchmarking code and documentation
  - Updated general documentation
