+++ benches/expand_from_coeff.rs
@@ -4,10 +4,14 @@ use whir::{crypto::fields::Field64, ntt};
 #[global_allocator]
 static ALLOC: AllocProfiler = AllocProfiler::system();
 
-// Test cases with polynomial sizes defined as exponents of 2 and expansion factors
+//
+// Test cases with tuple entries:
+//      - polynomial sizes defined as exponents of 2,
+//      - RS code expansion factors, and
+//      - interleaved bloc size exponent of 2
+//
 const TEST_CASES: &[(u32, usize, usize)] = &[
     (16, 2, 2),
-    (18, 2, 2),
     (18, 2, 2),
     (20, 2, 3),
     (16, 4, 3),
@@ -17,7 +21,7 @@ const TEST_CASES: &[(u32, usize, usize)] = &[
 ];
 
 #[divan::bench(args = TEST_CASES)]
-fn rs_encode_coset_poly(bencher: Bencher, case: &(u32, usize, usize)) {
+fn interleaved_rs_encode(bencher: Bencher, case: &(u32, usize, usize)) {
     bencher
         .with_inputs(|| {
             let (exp, expansion, coset_sz) = *case;
@@ -27,7 +31,7 @@ fn rs_encode_coset_poly(bencher: Bencher, case: &(u32, usize, usize)) {
             (coeffs, expansion, coset_sz)
         })
         .bench_values(|(coeffs, expansion, coset_sz)| {
-            black_box(ntt::rs_encode_coset_poly(&coeffs, expansion, coset_sz))
+            black_box(ntt::interleaved_rs_encode(&coeffs, expansion, coset_sz))
         });
 }
