+++ src/ntt/mod.rs
@@ -2,6 +2,10 @@
 
 mod cooley_tukey;
 mod matrix;
+
+#[cfg(test)]
+pub(crate) mod test_utils;
+
 mod transpose;
 mod utils;
 mod wavelet;
@@ -18,56 +22,47 @@ pub use self::{
 };
 
 ///
-/// RS encode a polynomial `poly` of degree `d-1` at the rate
-/// 1/`expansion` (i.e., the RS block length `N = d*expansion`).
+/// RS encode interleaved data `interleaved_coeffs` at the rate
+/// 1/`expansion`, where 2^`fold_factor` elements are interleaved
+/// together.
 ///
-/// Suppose the Merkle tree leaves are grouped as cosets of size  k =
-/// 2^`fold_factor`. Given a polynomial f(x) of degree `d-1`, it can be
-/// extended by zero-padding to a polynomial of degree `N-1` and written
-/// in the following form:
-///
-/// f(X) = h₀(X^k) + X·h₁(X^k) + ⋯ + X^{k-1}·h_{k-1}(X^k)
-///
-/// where each hₜ(X) is polynomial of maximum degree `N/k - 1`.
-///
-/// If ζ is the N-th root of unity, then [rs_encode_coset_poly] computes
-/// the Reed-Solomon code for each hₜ(x) using `ω = ζ^k` as the root of
-/// unity, and for each index i, places [h₀(ω^i), h₁(ω^i), ⋯
-/// ,h_{k-1}(ω^i)] consecutively in the output buffer.
+/// This function computes the RS-code for each interleaved message and
+/// outputs the interleaved alphabets in the same order as the input.
 ///
 #[cfg_attr(feature = "tracing", instrument(skip(poly), fields(size = poly.len())))]
-pub fn rs_encode_coset_poly<F: FftField>(
-    poly: &[F],
+pub fn interleaved_rs_encode<F: FftField>(
+    interleaved_coeffs: &[F],
     expansion: usize,
     fold_factor: usize,
 ) -> Vec<F> {
     let fold_factor = u32::try_from(fold_factor).unwrap();
     debug_assert!(expansion > 0);
-    debug_assert!(poly.len().is_power_of_two());
+    debug_assert!(interleaved_coeffs.len().is_power_of_two());
 
     let fold_factor_exp = 2usize.pow(fold_factor);
-    let expanded_size = poly.len() * expansion;
+    let expanded_size = interleaved_coeffs.len() * expansion;
 
     debug_assert_eq!(expanded_size % fold_factor_exp, 0);
 
-    // 1. Create zero-padded polynomial of appropriate size
+    // 1. Create zero-padded message of appropriate size
     let mut result = vec![F::zero(); expanded_size];
-    result[..poly.len()].copy_from_slice(poly);
+    result[..interleaved_coeffs.len()].copy_from_slice(interleaved_coeffs);
 
     let rows = expanded_size / fold_factor_exp;
     let columns = fold_factor_exp;
 
-    // 2. Create fold-factor-exp number of h(X^k) polynomial. (This is
-    //    equivalent to collecting the coefficients of `poly` in
-    //    multiples of `k = fold_factor_exp`). This is essentially
-    //    equivalent to transposing the matrix.
+    //
+    // 2. Convert from column-major (interleaved form) to row-major
+    //    representation.
+    //
+
+    // TODO: Might be useful to keep the transposed data for future use.
     transpose(&mut result, rows, columns);
 
-    // 3. Compute NTT for each hₜ(X), which will naturally be evaluated
-    //    at ω = ζ^k roots of unity.
+    // 3. Compute NTT on row-major representation
     ntt_batch(&mut result, rows);
 
-    // 4. Arrange cosets consecutively.
+    // 4. Convert back to column-major (interleaved) representation
     transpose(&mut result, columns, rows);
     result
 }
@@ -75,52 +70,9 @@ pub fn rs_encode_coset_poly<F: FftField>(
 #[cfg(test)]
 mod tests {
     use ark_ff::Field;
-    #[cfg(feature = "parallel")]
-    use rayon::prelude::*;
 
     use super::*;
     use crate::{crypto::fields::Field64, ntt::cooley_tukey::NttEngine};
-
-    fn expand_from_coeff<F: FftField>(coeffs: &[F], expansion: usize) -> Vec<F> {
-        let engine = cooley_tukey::NttEngine::<F>::new_from_cache();
-        let expanded_size = coeffs.len() * expansion;
-        let mut result = Vec::with_capacity(expanded_size);
-        // Note: We can also zero-extend the coefficients and do a larger NTT.
-        // But this is more efficient.
-
-        // Do coset NTT.
-        let root = engine.root(expanded_size);
-        result.extend_from_slice(coeffs);
-        #[cfg(not(feature = "parallel"))]
-        for i in 1..expansion {
-            let root = root.pow([i as u64]);
-            let mut offset = F::ONE;
-            result.extend(coeffs.iter().map(|x| {
-                let val = *x * offset;
-                offset *= root;
-                val
-            }));
-        }
-        #[cfg(feature = "parallel")]
-        result.par_extend((1..expansion).into_par_iter().flat_map(|i| {
-            let root_i = root.pow([i as u64]);
-            coeffs
-                .par_iter()
-                .enumerate()
-                .map_with(F::ZERO, move |root_j, (j, coeff)| {
-                    if root_j.is_zero() {
-                        *root_j = root_i.pow([j as u64]);
-                    } else {
-                        *root_j *= root_i;
-                    }
-                    *coeff * *root_j
-                })
-        }));
-
-        ntt_batch(&mut result, coeffs.len());
-        transpose(&mut result, expansion, coeffs.len());
-        result
-    }
 
     #[test]
     fn test_expand_from_coeff_size_2() {
@@ -174,7 +126,7 @@ mod tests {
         // The expected NTT result should be in transposed order:
         let expected_values_transposed = vec![expected_f0, expected_f2, expected_f1, expected_f3];
 
-        let computed_values = expand_from_coeff(&coeffs, expansion);
+        let computed_values = test_utils::expand_from_coeff(&coeffs, expansion);
         assert_eq!(computed_values, expected_values_transposed);
     }
 
@@ -290,16 +242,15 @@ mod tests {
             expected_f15,
         ];
 
-        let computed_values = expand_from_coeff(&coeffs, expansion);
+        let computed_values = test_utils::expand_from_coeff(&coeffs, expansion);
         assert_eq!(computed_values, expected_values_transposed);
     }
 
     #[test]
-    fn test_rs_encode_as_cosets() {
+    fn test_interleaved_rs_encode() {
         use ark_poly::{EvaluationDomain, GeneralEvaluationDomain};
         use ark_std::UniformRand;
 
-        use crate::poly_utils::fold::tests::transform_evaluations;
         let mut rng = ark_std::test_rng();
         let count = 1 << 20;
         let expansion = 4;
@@ -310,11 +261,15 @@ mod tests {
         let poly = Vec::from_iter((0..count).into_iter().map(|_| Field64::rand(&mut rng)));
 
         // Compute things the old way
-        let mut expected = expand_from_coeff(&poly, expansion);
-        transform_evaluations(&mut expected, eval_domain.group_gen_inv(), folding_factor);
+        let mut expected = test_utils::expand_from_coeff(&poly, expansion);
+        test_utils::transform_evaluations(
+            &mut expected,
+            eval_domain.group_gen_inv(),
+            folding_factor,
+        );
 
         // Compute things the new way
-        let interleaved_ntt = rs_encode_coset_poly(&poly, expansion, folding_factor);
+        let interleaved_ntt = interleaved_rs_encode(&poly, expansion, folding_factor);
         assert_eq!(expected, interleaved_ntt);
     }
 }
