+++ src/whir/prover.rs
@@ -19,7 +19,7 @@ use super::{
 };
 use crate::{
     domain::Domain,
-    ntt::rs_encode_coset_poly,
+    ntt::interleaved_rs_encode,
     poly_utils::{coeffs::CoefficientList, multilinear::MultilinearPoint},
     sumcheck::SumcheckSingle,
     utils::expand_randomness,
@@ -207,7 +207,7 @@ where
         let new_domain = round_state.domain.scale(2);
         let expansion = new_domain.size() / folded_coefficients.num_coeffs();
         let evals =
-            rs_encode_coset_poly(folded_coefficients.coeffs(), expansion, folding_factor_next);
+            interleaved_rs_encode(folded_coefficients.coeffs(), expansion, folding_factor_next);
 
         #[cfg(not(feature = "parallel"))]
         let leafs_iter = evals.chunks_exact(1 << folding_factor_next);
